2-26 修改migrations跟database以符合新的項目結構

+++ Core/Entities/Product.cs
@@ -12,8 +12,8 @@ namespace Core.Entities
         public decimal Price { get; set; }
         public string PictureUrl { get; set; }
         public ProductType ProductType { get; set; }
-        public int ProductTypeId { get; set; }
+        public long ProductTypeId { get; set; }
         public ProductBrand ProductBrand { get; set; }
-        public int ProductBrandId { get; set; }
+        public long ProductBrandId { get; set; }
     }
 }
